Replaced i/o table for [C000..C0FF] with table for [C000..CFFF] - New table uses units of 16-byte blocks - Each slot's card registers r/w handlers for C0xx & CXxx memory (slot 0-7) - TODO: Split ROMs so that Disk][ & SSC firmware is independent of main Apple ROM - TODO: Support $CFFF access & IORead_Cxxx() w/SLOTCXROM switch Also fixed reset bug where PC was read from reset vector before ROM had been switch back in

+++ AppleWin/source/Keyboard.cpp
@@ -314,7 +314,7 @@ static char ClipboardCurrChar(bool bIncPtr)
 
 //===========================================================================
 
-BYTE __stdcall KeybReadData (WORD, BYTE, BYTE, BYTE, ULONG)
+BYTE __stdcall KeybReadData (WORD, WORD, BYTE, BYTE, ULONG)
 {
 	keyboardqueries++;
 
@@ -352,7 +352,7 @@ BYTE __stdcall KeybReadData (WORD, BYTE, BYTE, BYTE, ULONG)
 
 //===========================================================================
 
-BYTE __stdcall KeybReadFlag (WORD, BYTE, BYTE, BYTE, ULONG)
+BYTE __stdcall KeybReadFlag (WORD, WORD, BYTE, BYTE, ULONG)
 {
 	keyboardqueries++;
